build: modularize isp
