+++ solid/src/main/java/solid/isp/RestAction_Hierarchy.puml
@@ -1,0 +1,55 @@
+@startuml
+
+interface RestAction {
+    +onRest(): String
+}
+
+interface ChatCapable extends RestAction {
+    +chat(): String
+}
+
+interface ResetCapable extends RestAction {
+    +reset(): String
+}
+
+interface ChatResetCapable extends ChatCapable, ResetCapable {
+}
+
+class MorningChat {
+    +onRest(): String
+    +chat(): String
+}
+
+class HelloChat {
+    +onRest(): String
+    +chat(): String
+}
+
+class QuietReset {
+    +onRest(): String
+    +reset(): String
+}
+
+class NoisyReset {
+    +onRest(): String
+    +reset(): String
+}
+
+class DefaultChatResetCapable {
+    +chat(): String
+    +reset(): String
+    +onRest(): String
+}
+
+RestAction <|-- ChatCapable
+RestAction <|-- ResetCapable
+ChatCapable <|-- ChatResetCapable
+ResetCapable <|-- ChatResetCapable
+
+ChatCapable <|-- MorningChat
+ChatCapable <|-- HelloChat
+ResetCapable <|-- QuietReset
+ResetCapable <|-- NoisyReset
+ChatResetCapable <|-- DefaultChatResetCapable
+
+@enduml
